Fixed CPU crash
The crash was caused by the acia generating a FIRQ.
The CPU was executing FIRQ and then crashed.
The reason is that the current model of the acia reads from Serial. Basically it reads garbage, causing the FIRQ to fire.
The ACIA should not fire a FIRQ until we send MIDI data.
I commented the Serial.read in acia.cpp to make this work.
We will connect to MIDI next.