Dec 12th update
Disabled the "read" from ACIA, it is causing the CPU to crash - we will investigate later.

Now the system gets stuck in a loop due to the T2 Timer (we just enter the IRQ routine when T2 fires, to update T2 again, and so on).

There is something we do not understand regarding how to service the IRQ likely the VIA one.
We know it is not from the DOC, because the DOC has been initialized with all interrupts DISABLED.

Progress on DOC.

More comments on BOOT and OS3.2 ROMs